Support micro-images (scratch) by static linking
CGO_ENABLED=1 is the default and [enables the creation of go code that calls C code](https://golang.org/cmd/cgo/#hdr-Special_cases) – but the wrinkle is that the binary produced is not statically linked. This causes issues we try to insert the built binary in a `scratch` or `alpine` docker microimage.